requirements.txt
streamlit
langchain
chromadb
faiss-cpu
openai
pypdf
python-pptx
sentence-transformers
tqdm
llama-cpp-python
watchdog